fix: move data migration notice from default layout to /app page
Show the v0.1.1 data isolation upgrade notice only when users
navigate to /app, instead of on every page using the default layout.
